+++ drasyl-cli/src/main/java/org/drasyl/cli/rc/handler/JsonRpc2RequestHandler.java
@@ -23,15 +23,24 @@ package org.drasyl.cli.rc.handler;
 
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.SimpleChannelInboundHandler;
+import org.drasyl.cli.util.IdentityPublicKeyMixin;
+import org.drasyl.cli.util.PeerMixin;
+import org.drasyl.cli.util.PeersListMixin;
+import org.drasyl.cli.util.RoleMixin;
 import org.drasyl.cli.node.message.JsonRpc2Error;
 import org.drasyl.cli.node.message.JsonRpc2Request;
 import org.drasyl.cli.node.message.JsonRpc2Response;
+import org.drasyl.handler.peers.Peer;
+import org.drasyl.handler.peers.PeersList;
+import org.drasyl.handler.peers.Role;
 import org.drasyl.identity.Identity;
+import org.drasyl.identity.IdentityPublicKey;
 
 import java.util.Map;
 
 import static io.netty.channel.ChannelFutureListener.FIRE_EXCEPTION_ON_FAILURE;
 import static org.drasyl.cli.node.message.JsonRpc2Error.METHOD_NOT_FOUND;
+import static org.drasyl.node.JsonUtil.JACKSON_MAPPER;
 
 public abstract class JsonRpc2RequestHandler extends SimpleChannelInboundHandler<JsonRpc2Request> {
     protected void requestMethodNotFound(final ChannelHandlerContext ctx,
@@ -55,5 +64,14 @@ public abstract class JsonRpc2RequestHandler extends SimpleChannelInboundHandler
                         "secretKey", identity.getKeyAgreementSecretKey().toUnmaskedString()
                 )
         );
+    }
+
+    protected Map<String, Object> peersMap(final PeersList peers) {
+        JACKSON_MAPPER.addMixIn(PeersList.class, PeersListMixin.class);
+        JACKSON_MAPPER.addMixIn(IdentityPublicKey.class, IdentityPublicKeyMixin.class);
+        JACKSON_MAPPER.addMixIn(Peer.class, PeerMixin.class);
+        JACKSON_MAPPER.addMixIn(Role.class, RoleMixin.class);
+
+        return JACKSON_MAPPER.convertValue(peers, Map.class);
     }
 }